chore: rename fastCheckout to checkout
BREAKING CHANGE: the `fastCheckout` function has been renamed `checkout` and the old `checkout` function has been removed entirely.
