+++ __tests__/test-submodules.js
@@ -5,7 +5,6 @@ const {
   clone,
   checkout,
   listFiles,
-  fastCheckout,
   commit
 } = require('isomorphic-git')
 
@@ -50,7 +49,7 @@ describe('submodule "support"', () => {
     expect(await listFiles({ gitdir })).toContain('test.empty')
   })
 
-  it('submodules are staged when switching to a branch that has them (checkout)', async () => {
+  it('submodules are staged when switching to a branch that has them', async () => {
     const { dir, gitdir } = await makeFixture('test-clone-submodules')
     await clone({
       dir,
@@ -71,7 +70,7 @@ describe('submodule "support"', () => {
     expect(await listFiles({ gitdir })).toContain('test.empty')
   })
 
-  it("submodules are unstaged when switching to a branch that doesn't have them (checkout)", async () => {
+  it("submodules are unstaged when switching to a branch that doesn't have them", async () => {
     const { dir, gitdir } = await makeFixture('test-clone-submodules')
     await clone({
       dir,
@@ -82,41 +81,6 @@ describe('submodule "support"', () => {
     })
     // Test
     await checkout({ dir, gitdir, ref: 'no-modules' })
-    expect(await listFiles({ gitdir })).not.toContain('test.empty')
-  })
-
-  it('submodules are staged when switching to a branch that has them (fastCheckout)', async () => {
-    const { dir, gitdir } = await makeFixture('test-clone-submodules')
-    await clone({
-      dir,
-      gitdir,
-      ref: 'no-modules',
-      url: `http://${localhost}:8888/test-submodules.git`,
-      noSubmodules: true,
-      newSubmoduleBehavior: true
-    })
-    // Test
-    await fastCheckout({
-      dir,
-      gitdir,
-      ref: 'master',
-      noSubmodules: true,
-      newSubmoduleBehavior: true
-    })
-    expect(await listFiles({ gitdir })).toContain('test.empty')
-  })
-
-  it("submodules are unstaged when switching to a branch that doesn't have them (fastCheckout)", async () => {
-    const { dir, gitdir } = await makeFixture('test-clone-submodules')
-    await clone({
-      dir,
-      gitdir,
-      url: `http://${localhost}:8888/test-submodules.git`,
-      noSubmodules: true,
-      newSubmoduleBehavior: true
-    })
-    // Test
-    await fastCheckout({ dir, gitdir, ref: 'no-modules' })
     expect(await listFiles({ gitdir })).not.toContain('test.empty')
   })
 })
