+++ src/index.d.ts
@@ -372,9 +372,13 @@ export function checkout(args: WorkDir & GitDir & {
   emitter?: EventEmitter;
   emitterPrefix?: string;
   remote?: string;
-  ref: string;
+  ref?: string;
   filepaths?: string[];
-  pattern?: string;
+  dryRun?: boolean;
+  debug?: boolean;
+  force?: boolean;
+  noCheckout?: boolean;
+  noUpdateHead?: boolean;
   noSubmodules?: boolean;
   newSubmoduleBehavior?: boolean;
 }): Promise<void>;
@@ -481,23 +485,6 @@ export function expandOid(args: GitDir & {
   fs?: any;
   oid: string;
 }): Promise<string>;
-
-export function fastCheckout(args: WorkDir & GitDir & {
-  core?: string;
-  fs?: any;
-  emitter?: EventEmitter;
-  emitterPrefix?: string;
-  remote?: string;
-  ref?: string;
-  filepaths?: string[];
-  dryRun?: boolean;
-  debug?: boolean;
-  force?: boolean;
-  noCheckout?: boolean;
-  noUpdateHead?: boolean;
-  noSubmodules?: boolean;
-  newSubmoduleBehavior?: boolean;
-}): Promise<void>;
 
 export function fetch(args: GitDir & {
   core?: string;
