+++ src/index.js
@@ -7,7 +7,6 @@ export * from './commands/addRemote.js'
 export * from './commands/annotatedTag.js'
 export * from './commands/branch.js'
 export * from './commands/checkout.js'
-export * from './commands/fastCheckout.js'
 export * from './commands/clone.js'
 export * from './commands/commit.js'
 export * from './commands/config.js'
