+++ website/sidebars.json
@@ -12,7 +12,6 @@
       "config",
       "fetch",
       "checkout",
-      "fastCheckout",
       "push",
       "pull",
       "merge",
